# Improve LocalNodeStore.Get() performance and fix possible deadlock