--- a/pkg/node/local_node_store.go
+++ b/pkg/node/local_node_store.go
@@ -5,6 +5,7 @@ package node
 
 import (
 	"context"
+	"io"
 	"sync"
 
 	"github.com/cilium/stream"
@@ -58,8 +59,23 @@ type LocalNodeStore struct {
 	// Changes to the local node are observable.
 	stream.Observable[LocalNode]
 
-	mu       lock.Mutex
+	// mu is the main LocalNodeStore mutex, which protects the access to the
+	// different fields during all operations. getMu, instead, is a separate
+	// mutex which is used to guard updates of the value field, as well as its
+	// access by the Get() method. The reason for using two separate mutexes
+	// being that we don't want Get() to be blocked while calling emit, as
+	// that synchronously calls into all subscribers, which is a potentially
+	// expensive operation, and a possible source of deadlocks (e.g., one of
+	// the subscribers needs to acquire another mutex, which is held by a
+	// separate goroutine trying to call LocalNodeStore.Get()). In addition,
+	// getMu also guards the complete field, as it is used by Get() to
+	// determine that the LocalNodeStore was stopped. When both mu and getMu
+	// are to be acquired together, mu shall be always acquired first.
+	mu    lock.Mutex
+	getMu lock.RWMutex
+
 	value    LocalNode
+	hasValue <-chan struct{}
 	emit     func(LocalNode)
 	complete func(error)
 }
@@ -72,11 +88,17 @@ func NewTestLocalNodeStore(mockNode LocalNode) *LocalNodeStore {
 		emit:       emit,
 		complete:   complete,
 		value:      mockNode,
+		hasValue: func() <-chan struct{} {
+			ch := make(chan struct{})
+			close(ch)
+			return ch
+		}(),
 	}
 }
 
 func NewLocalNodeStore(params LocalNodeStoreParams) (*LocalNodeStore, error) {
 	src, emit, complete := stream.Multicast[LocalNode](stream.EmitLatest)
+	hasValue := make(chan struct{})
 
 	s := &LocalNodeStore{
 		Observable: src,
@@ -86,6 +108,7 @@ func NewLocalNodeStore(params LocalNodeStoreParams) (*LocalNodeStore, error) {
 			Labels:      make(map[string]string),
 			Annotations: make(map[string]string),
 		}},
+		hasValue: hasValue,
 	}
 
 	bctx, cancel := context.WithCancel(context.Background())
@@ -116,6 +139,7 @@ func NewLocalNodeStore(params LocalNodeStoreParams) (*LocalNodeStore, error) {
 			s.emit = emit
 			s.complete = complete
 			emit(s.value)
+			close(hasValue)
 			return nil
 		},
 		OnStop: func(cell.HookContext) error {
@@ -125,8 +149,10 @@ func NewLocalNodeStore(params LocalNodeStoreParams) (*LocalNodeStore, error) {
 
 			s.mu.Lock()
 			s.complete(nil)
+			s.getMu.Lock()
 			s.complete = nil
 			s.emit = nil
+			s.getMu.Unlock()
 			s.mu.Unlock()
 
 			localNode = nil
@@ -141,17 +167,34 @@ func NewLocalNodeStore(params LocalNodeStoreParams) (*LocalNodeStore, error) {
 // e.g. in API handlers. Do not assume the value does not change over time.
 // Blocks until the store has been initialized.
 func (s *LocalNodeStore) Get(ctx context.Context) (LocalNode, error) {
-	// Subscribe to the stream of updates and take the first (latest) state.
-	return stream.First[LocalNode](ctx, s)
+	select {
+	case <-s.hasValue:
+		s.getMu.RLock()
+		defer s.getMu.RUnlock()
+
+		if s.complete == nil {
+			// Return EOF when the LocalNodeStore is stopped, to preserve the
+			// same behavior of stream.First[LocalNode].
+			return LocalNode{}, io.EOF
+		}
+
+		return s.value, nil
+
+	case <-ctx.Done():
+		return LocalNode{}, ctx.Err()
+	}
 }
 
 // Update modifies the local node with a mutator. The updated value
-// is passed to observers.
+// is passed to observers. Calling LocalNodeStore.Get() from the
+// mutation function is forbidden, and would result in a deadlock.
 func (s *LocalNodeStore) Update(update func(*LocalNode)) {
 	s.mu.Lock()
 	defer s.mu.Unlock()
 
+	s.getMu.Lock()
 	update(&s.value)
+	s.getMu.Unlock()
 
 	if s.emit != nil {
 		s.emit(s.value)