Currently, the local node store is implemented on top of an observable stream, and the Get operation leverages stream.First to retrieve the last emitted value. However, this approach suffers from two drawbacks, further detailed below: performance overhead and risk of deadlocks. To address them, let's instead implement the Get() method so that it directly returns a shallow copy of the stored LocalNode object, while making sure to preserve the existing behavior before starting the LocalNodeStore, and after stopping it.

More in detail, the possible deadlock is caused by the following series of events:

1. LocalNodeStore.Update() is triggered. stream.Multicast acquires a lock, and then synchronously calls the subscribers. Subscribers may do arbitrarily complex operations at that point, possibly acquiring other locks in the meanwhile. Let's assume that one of them blocks waiting for a given lock mu. 
2. A separate thread is currently holding the mu lock, and calls one of the global getters to, let's say, retrieve the v4 NodeInternalIP address, which in turns calls LocalNodeStore.Get() and creates a new stream.First observing the LocalNodeStore updates. However, the replay of the last event is blocked by the stream.Multicast lock held by the other thread, effectively causing a deadlock.

The subtle aspect here is that the global getters appear innocuous, and it is difficult to reason about the possibility of deadlocks. Conversely, the new approach ensures that LocalNodeStore.Get() never blocks even if any subscriber is still processing the update event.

In terms of performance, taking into account the newly introduced BenchmarkLocalNodeStoreGet benchmark, the new implementation cuts the CPU time and memory allocations by 96% and 100% respectively:

```
goos: linux
goarch: amd64
pkg: github.com/cilium/cilium/pkg/node
cpu: 12th Gen Intel(R) Core(TM) i7-12700H
                     │    old.txt    │               new.txt                 │
                     │    sec/op     │   sec/op     vs base                  │
LocalNodeStoreGet-20   1151.00n ± 1%   44.27n ± 1%  -96.15% (p=0.000 n=10)

                     │   old.txt     │                new.txt                │
                     │     B/op      │     B/op      vs base                 │
LocalNodeStoreGet-20   1.090Ki ± 0%    0.000Ki ± 0%  -100.00% (p=0.000 n=10)

                     │  old.txt      │              new.txt                  │
                     │ allocs/op     │ allocs/op  vs base                    │
LocalNodeStoreGet-20   13.00 ± 0%      0.00 ± 0%  -100.00% (p=0.000 n=10)
```

Note: I've currently marked the PR as `release-note/misc`, and not marked it for backport given that the discovered deadlock also involved https://github.com/cilium/cilium/pull/30470, which got merged only recently. Happy to reconsider that decision if reviewers think that it is worth to be backported.